Convert eager-eval utils to actions
