+++ src/devtools/client/webconsole/actions/autocomplete-eager.ts
@@ -1,0 +1,57 @@
+import { getSelectedFrame } from "devtools/client/debugger/src/selectors";
+
+import { ThreadFront, ValueFront } from "protocol/thread";
+import { getPropertiesForObject } from "../utils/autocompleteomplete";
+
+// Use eager eval to get the properties of the last complete object in the expression.
+export async function getEvaluatedProperties(expression: string): Promise<string[]> {
+  return async ({ getState }) => {
+    const { asyncIndex, frameId } = getSelectedFrame(getState());
+
+    try {
+      const { returned, exception, failed } = await ThreadFront.evaluate({
+        asyncIndex,
+        frameId,
+        text: expression,
+        pure: true,
+      });
+      if (returned && !(failed || exception)) {
+        return getPropertiesForObject(returned.getObject());
+      }
+    } catch (err: any) {
+      let msg = "Error: Eager Evaluation failed";
+      if (err.message) {
+        msg += ` - ${err.message}`;
+      }
+      console.error(msg);
+    }
+
+    return [];
+  };
+}
+
+export async function eagerEvaluateExpression(expression: string): Promise<ValueFront | null> {
+  return async ({ getState }) => {
+    const { asyncIndex, frameId } = getSelectedFrame(getState());
+
+    try {
+      const { returned, exception, failed } = await ThreadFront.evaluate({
+        asyncIndex,
+        frameId,
+        text: expression,
+        pure: true,
+      });
+      if (returned && !(failed || exception)) {
+        return returned;
+      }
+    } catch (err: any) {
+      let msg = "Error: Eager Evaluation failed";
+      if (err.message) {
+        msg += ` - ${err.message}`;
+      }
+      console.error(msg);
+    }
+
+    return null;
+  };
+}
