--- src/devtools/client/webconsole/utils/autocomplete-eager.ts
@@ -1,51 +1,0 @@
-import { ThreadFront, ValueFront } from "protocol/thread";
-import { getPropertiesForObject } from "./autocomplete";
-
-// Use eager eval to get the properties of the last complete object in the expression.
-export async function getEvaluatedProperties(expression: string): Promise<string[]> {
-  const { asyncIndex, frameId } = gToolbox.getPanel("debugger")!.getFrameId();
-
-  try {
-    const { returned, exception, failed } = await ThreadFront.evaluate({
-      asyncIndex,
-      frameId,
-      text: expression,
-      pure: true,
-    });
-    if (returned && !(failed || exception)) {
-      return getPropertiesForObject(returned.getObject());
-    }
-  } catch (err: any) {
-    let msg = "Error: Eager Evaluation failed";
-    if (err.message) {
-      msg += ` - ${err.message}`;
-    }
-    console.error(msg);
-  }
-
-  return [];
-}
-
-export async function eagerEvaluateExpression(expression: string): Promise<ValueFront | null> {
-  const { asyncIndex, frameId } = gToolbox.getPanel("debugger")!.getFrameId();
-
-  try {
-    const { returned, exception, failed } = await ThreadFront.evaluate({
-      asyncIndex,
-      frameId,
-      text: expression,
-      pure: true,
-    });
-    if (returned && !(failed || exception)) {
-      return returned;
-    }
-  } catch (err: any) {
-    let msg = "Error: Eager Evaluation failed";
-    if (err.message) {
-      msg += ` - ${err.message}`;
-    }
-    console.error(msg);
-  }
-
-  return null;
-}
